[vibe] follow ups should work -_-
